# Utilize multithreading and asynchronous operations