--- a/backend/requirements.txt
+++ b/backend/requirements.txt
@@ -7,3 +7,6 @@ mysqlclient
 django-cors-headers
 djangorestframework
 pymysql
+aiohttp
+aiodns
+cryptography