--- a/backend/scraperapp/.gitignore
+++ b/backend/scraperapp/.gitignore
@@ -1,0 +1,1 @@
+html_data/*