--- a/backend/scraperapp/constants.py
+++ b/backend/scraperapp/constants.py
@@ -1,9 +1,14 @@
 # URLs
-CATALOG_URL = 'https://catalog.ucdavis.edu/home/'
+CATALOG_URL = 'https://catalog.ucdavis.edu/courses-subject-code/'
 SUBJECT_URL = 'https://catalog.ucdavis.edu/courses-subject-code/{}/'
 
 # Database handling
 IGNORE_SUBJECTS = [
     'MMG', # Courses coming soon
     'DVM', # Courses are not displayed
+    'PDF',
 ]
+
+# HTML Path
+DATA_DIR = 'html_data'
+SUBJECT_DIR = 'subjects'