--- a/backend/scraperapp/processor.py
+++ b/backend/scraperapp/processor.py
@@ -16,13 +16,17 @@ def process_catalog(catalog_file):
     soup = create_soup(catalog_file)
 
     subjects = soup.find_all('a', {'href': re.compile(r'^/courses-subject-code/[a-z]+')})
+    length = len(subjects)
+    ignore_count = 0
 
-    for subject in subjects:
+    for i, subject in enumerate(subjects):
+        print(f'\rProcessed {i}/{length} subjects', end='')
         subject.text.replace('\u200b', '')
         subject_name, subject_code = re.search(r'(.+) \(([A-Z]{3})\)', subject.text).groups()
 
         if subject_code in IGNORE_SUBJECTS:
-            print(f'Ignoring {subject_code}')
+            # print(f'Ignoring {subject_code}')
+            ignore_count += 1
             continue
 
         try:
@@ -31,42 +35,79 @@ def process_catalog(catalog_file):
             setattr(obj, 'code', subject_code)
             setattr(obj, 'name', subject_name)
             obj.save()
-            print(f'Entry {subject_code} Updated')
+            # print(f'Entry {subject_code} Updated')
         except Subject.DoesNotExist:
             obj = Subject(code=subject_code, name=subject_name)
             obj.save()
-            print(f'Entry {subject_code} Saved')
+            # print(f'Entry {subject_code} Saved')
+    print(f'\rProcessed {length}/{length} subjects. Ignored {ignore_count}.')
 
-# def process_courses(subject_html, subject_obj):
+def process_subjects(subject_paths):
+    length = len(subject_paths)
+    course_count = 0
+
+    for i, (code, filepath) in enumerate(subject_paths):
+        print(f'\rProcessed {i}/{length} subject courses. Total count: {course_count}', end='')
+        course_count += process_subject(code, filepath)
+
+    print(f'\rProcessed {length}/{length} subject courses. Total count: {course_count}')
+
+def process_subject(subject_code, filepath):
+
+    soup = create_soup(filepath)
+    courses = soup.find_all('div', {'class': 'courseblock'})
+    obj = Subject.objects.get(code=subject_code)
+
+    for course in courses:
+        process_course(course, obj)
+
+    return len(courses)
+
+def process_course(course_html, subject_obj):
+
+    soup = bs(str(course_html), 'html.parser')
+    course = dict()
+    course['code'] = soup.find('span', {'class': 'detail-code'}).text.strip()
     
-#     for course in subject_html:
-#         soup = bs(str(course), 'html.parser')
-#         course_code = soup.find('span', {'class': 'detail-code'}).text.strip()
-        
-#         # Skip objects already in database
-#         if Course.objects.filter(code=course_code).exists():
-#             continue
-        
-#         title = soup.find('span', {'class': 'detail-title'}).text.strip()[2:]
-#         description = soup.find('p', {'class': 'courseblockextra noindent'}).text.strip()
-#         description = description.removeprefix('Course Description: ')
-        
-#         prerequisites = soup.find('p', {'class': 'detail-prerequisite'})
-#         if prerequisites is not None:
-#             prerequisites = prerequisites.text.strip()
-#             prerequisites = prerequisites.removeprefix('Prerequisite(s): ')
-#             prerequisites = prerequisites.removesuffix('.')
-#             prerequisites = prerequisites.replace('\xa0', ' ')
-            
-#         course_object = Course(
-#             code=course_code,
-#             title=title,
-#             subject=subject_obj,
-#             description=description,
-#             prerequisites=prerequisites
-#         )
-        
-#         course_object.save()
+    course['title'] = soup.find('span', {'class': 'detail-title'}).text.strip()[2:]
+    description = soup.find('p', {'class': 'courseblockextra noindent'}).text.strip()
+    course['description'] = description.removeprefix('Course Description: ')
+    
+    prerequisites = soup.find('p', {'class': 'detail-prerequisite'})
+    if prerequisites is not None:
+        prerequisites = prerequisites.text.strip()
+        prerequisites = prerequisites.removeprefix('Prerequisite(s): ')
+        prerequisites = prerequisites.removesuffix('.')
+        prerequisites = prerequisites.replace('\xa0', ' ')
+
+    course['prerequisites'] = prerequisites
+    course['subject'] = subject_obj
+
+    try:
+        obj = Course.objects.get(code=course['code'])
+
+        for key, val in course.items():
+            setattr(obj, key, val)
+
+        obj.save()
+
+    except Course.DoesNotExist:
+        obj = Course(
+            code=course['code'],
+            title=course['title'],
+            subject=course['subject'],
+            description=course['description'],
+            prerequisites=course['prerequisites']
+        )
+    
+        obj.save()
+
+def process_prerequisites(subjects):
+
+    for obj in subjects:
+        courses = Course.objects.filter(subject=obj, prerequisites__isnull=False)
+
+        # for course in courses:
 
 
 # def process_prerequisites(subject_obj):