… backend/scraperapp/migrations/0001_initial.py
… backend/scraperapp/migrations/0002_prerequisite_subject_id.py
… backend/scraperapp/migrations/0003_prerequisite_course_code_alter_course_subject_and_more.py
… backend/scraperapp/scraper.py
… backend/scraperapp/web_crawler.py