Previously, all operations were done sequentially on a single thread. Thus running `python manage.py scrape` would require approximately 8 minutes to run on an M1 chip. This is because we need to process 219 courses, over 10,000 courses, and over 10,000 prerequisites. Utilizing asynchronous operations allows us to request multiple sites at the same time without waiting for a request to complete. Multithreading allows us process multiple courses or prerequisites at the same time and insert them all into the databse.

Together, these two optimizations allowed for the following speedups:

|                          | Before | After |
|--------------------------|--------|-------|
| Fetching Subject Pages   | 2m21s  | 7s    |
| Processing Courses       | 1m14s  | 30s   |
| Processing Prerequisites | 1m30s  | 12s   |